Set default user/group to nobody/nogroup
Default group used to be "#-1".  When trying this, it works out to
initgroups("nobody", -1) and setgroups(1, [4294967295]) which returned
EINVAL and crashes.  Best we can tell, -1 is a hack which used to work
way back when, but isn't a documented feature.

Setting the default to nobody/nogroup isn't very portable (did AIX use
nogroup or nobody???), but hey, most Linux distros use that.